+++ src/libsyntax/lib.rs
@@ -87,6 +87,7 @@ pub mod ext {
         pub mod macro_rules;
     }
 
+    pub mod mtwt;
 
     pub mod cfg;
     pub mod fmt;
